requirements.txt
streamlit==1.37.1
langchain==0.2.15
langchain-community==0.2.11
langchain-text-splitters==0.2.4
langchain-groq==0.1.5
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-dotenv==1.0.1
httpx==0.27.0
